+++ src/main/java/com/elbuensabor/apirest/Controller/AdminController.java
@@ -1,0 +1,4 @@
+package com.elbuensabor.apirest.Controller;
+
+public class AdminController {
+}
