+++ src/main/java/com/elbuensabor/apirest/Controller/UserController.java
@@ -1,0 +1,4 @@
+package com.elbuensabor.apirest.Controller;
+
+public class UserController {
+}
