--- src/main/java/com/elbuensabor/apirest/Controllers/UserController.java
@@ -1,4 +1,0 @@
-package com.elbuensabor.apirest.Controllers;
-
-public class UserController {
-}
