--- src/main/java/com/elbuensabor/apirest/Entities/User.java
@@ -1,4 +1,0 @@
-package com.elbuensabor.apirest.Entities;
-
-public class User {
-}
