+++ src/main/java/com/elbuensabor/apirest/Exception/MiException.java
@@ -1,0 +1,4 @@
+package com.elbuensabor.apirest.Exception;
+
+public class MiException {
+}
