--- src/main/java/com/elbuensabor/apirest/Exceptions/MiException.java
@@ -1,4 +1,0 @@
-package com.elbuensabor.apirest.Exceptions;
-
-public class MiException {
-}
