--- src/main/java/com/elbuensabor/apirest/Repositories/UserRepository.java
@@ -1,4 +1,0 @@
-package com.elbuensabor.apirest.Repositories;
-
-public class UserRepository {
-}
