+++ src/main/java/com/elbuensabor/apirest/Repository/UserRepository.java
@@ -1,0 +1,4 @@
+package com.elbuensabor.apirest.Repository;
+
+public class UserRepository {
+}
