+++ src/main/java/com/elbuensabor/apirest/Service/UserService.java
@@ -1,0 +1,4 @@
+package com.elbuensabor.apirest.Service;
+
+public class UserService {
+}
