--- src/main/java/com/elbuensabor/apirest/Services/UserService.java
@@ -1,4 +1,0 @@
-package com.elbuensabor.apirest.Services;
-
-public class UserService {
-}
